2nd GitHub Trial
Publishing this Stomp repository to make it easier to edit project files
this update is as of the 1.6.7 update on google drive.

Please don't save Windows or Mac builds on here, publish those directly
to google drive.  There are a few limitation to github involving file
size (1GB total, 100mb per file) and the playable builds are in excess
of the 100 mb cap and could push the total file size to over 1GB.